% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_unit.R
\name{cell.unit}
\alias{cell.unit}
\title{Identify metacells in single cell genomic data}
\usage{
cell.unit(
  object,
  reduction = "pca",
  dims = 50,
  step = 20,
  min.cells = 10,
  seed = 1024,
  Q.final = 0.75,
  Q.sub = 0.6
)
}
\arguments{
\item{object}{A seurat based object containing assays, reduction, meta data etc.}

\item{reduction}{The reduction slot for ICA analysis. Default is pca.}

\item{dims}{The target dimensions for ICA analysis. Default is 50.}

\item{steps}{The number of steps for diffusion condensation in the independent component (IC) space. The default value is set to 20.}

\item{min.cells}{The minimal number of cells in a cell cluster/subcluster. The default value is set to 10.}

\item{seed}{The seed for generating intial w.init for fastICA. Default is 1024.}

\item{Q.final}{The inner connectivity threshold for discarding metacells with less than 20 cells in the final refine step. The default value is set to 0.75.}

\item{Q.sub}{The inner connectivity threshold for evaluating over partition in the recursive step. The default value is set to 0.6.}
}
\value{
Seurat object with metacell grouping information stored in the metacell of meta.data slot in object
}
\description{
CellUnit aggregates cells into compact metacells based on cell condensation and recursive splitting with curated termination criteria.
}
\examples{
## Starts from counts
library(CellUnit)
library(Seurat)
sc_object <- CreateSeuratObject(count=counts, project = "sc_object", min.cells = 3)
sc_object$percent.mt <- PercentageFeatureSet(sc_object, pattern = "^MT-")
sc_object <- subset(sc_object, percent.mt<20)
sc_object <- NormalizeData(sc_object)
sc_object <- FindVariableFeatures(sc_object, nfeatures=2000)
sc_object <- ScaleData(sc_object, do.center = F)
sc_object <- RunPCA(sc_object, npcs=50)
sc_object <- FindNeighbors(object = sc_object,
                           k.param = 20,
                           compute.SNN = F,
                           prune.SNN = 0,
                           reduction = "pca",
                           dims = 1:50,
                           force.recalc = F, return.neighbor = T)
sc_object <- cell.unit(sc_object)

}
