% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLayers.R
\name{get.layers}
\alias{get.layers}
\title{Get cellular hierarchy based on the identified fine-grained cell clusters}
\usage{
get.layers(
  object
)
}
\arguments{
\item{object}{A seurat object processed by FindMetacells function.}
}
\value{
A matrix containing multi-resolution clusters by columns, with each row representing a cell.
}
\description{
get.layers infers cellular hierarchy based on the simlarity of aggregated expression of identified fine-grained clusters.
}
\examples{
## Starts from counts
library(scArch)
library(Seurat)
sc_object <- CreateSeuratObject(count=counts, project = "sc_object", min.cells = 3)
sc_object$percent.mt <- PercentageFeatureSet(sc_object, pattern = "^MT-")
sc_object <- subset(sc_object, percent.mt<20)
sc_object <- NormalizeData(sc_object)
sc_object <- FindVariableFeatures(sc_object, nfeatures=2000)
sc_object <- ScaleData(sc_object, do.center = F)
sc_object <- RunPCA(sc_object, npcs=50)
sc_object <- FindNeighbors(object = sc_object,
                           k.param = 20,
                           compute.SNN = F,
                           prune.SNN = 0,
                           reduction = "pca",
                           dims = 1:50,
                           force.recalc = F, return.neighbor = T)
sc_object <- FindMetacells(sc_object)

Layered.cluster = get.layers(object = sc_object)

}
